Prepare for maven central publishing
