+++ flink-connector-ethereum/pom.xml
@@ -23,7 +23,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
-        <groupId>org.kaigorodov.flink</groupId>
+        <groupId>io.github.kaygorodov</groupId>
         <artifactId>flink-connector-ethereum-parent</artifactId>
         <version>0.1.0-SNAPSHOT</version>
     </parent>
