+++ flink-connector-ethereum/src/main/java/io/github/kaygorodov/flink/connector/ethereum/EthereumEnumeratorState.java
@@ -1,0 +1,30 @@
+/*
+ * Copyright © 2024 Andrei Kaigorodov (andreykaygorodov@gmail.com)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.github.kaygorodov.flink.connector.ethereum;
+
+import java.math.BigInteger;
+
+public class EthereumEnumeratorState {
+  private final BigInteger lastAssignedBlockNumber;
+
+  public BigInteger getLastAssignedBlockNumber() {
+    return lastAssignedBlockNumber;
+  }
+
+  public EthereumEnumeratorState(BigInteger lastAssignedBlockNumber) {
+    this.lastAssignedBlockNumber = lastAssignedBlockNumber;
+  }
+}
