+++ pom.xml
@@ -20,10 +20,12 @@
         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
         xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>org.kaigorodov.flink</groupId>
+    <groupId>io.github.kaygorodov</groupId>
     <artifactId>flink-connector-ethereum-parent</artifactId>
     <version>0.1.0-SNAPSHOT</version>
     <name>Flink Ethereum Connector Parent</name>
+    <description>Flink DataStream connector for any Ethereum-compatible network</description>
+    <url>https://github.com/kaygorodov/flink-connector-ethereum</url>
     <packaging>pom</packaging>
     <inceptionYear>2024</inceptionYear>
     <parent>
@@ -31,6 +33,22 @@
         <artifactId>flink-connector-parent</artifactId>
         <version>1.0.0</version>
     </parent>
+
+    <licenses>
+        <license>
+            <name>The Apache License, Version 2.0</name>
+            <url>http://www.apache.org/licenses/LICENSE-2.0.txt</url>
+        </license>
+    </licenses>
+
+    <developers>
+        <developer>
+            <name>Andrei Kaigorodov</name>
+            <email>andreykaygorodov@gmail.com</email>
+            <organization>Andrei Kaigorodov</organization>
+            <organizationUrl>https://github.com/kaygorodov</organizationUrl>
+        </developer>
+    </developers>
 
     <scm>
         <url>https://github.com/kaygorodov/flink-connector-ethereum</url>
